@@ -1,6 +1,6 @@
 {
   "name": "react-activity",
-  "version": "1.0.16",
+  "version": "1.1.0",
   "description": "Activity indicator component library for React applications.",
   "main": "./dist/react-activity.js",
   "scripts": {
@@ -41,7 +41,7 @@
     "babel-preset-react": "^6.3.13",
     "css-loader": "^0.23.0",
     "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.11.3",
+    "eslint-plugin-react": "^3.16.1",
     "extract-text-webpack-plugin": "^0.9.1",
     "node-sass": "^3.4.2",
     "pre-commit": "^1.1.2",
@@ -51,8 +51,10 @@
     "webpack-dev-server": "^1.14.0"
   },
   "peerDependencies": {
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1"
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4"
   },
-  "dependencies": {}
+  "dependencies": {
+    "prop-types": "^15.5.8"
+  }
 }
