@@ -1,40 +1,36 @@
 'use strict';
-import React, {PropTypes} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 export default ((ComposedComponent, defaultAnimationDuration) => {
-  return React.createClass({
-    displayName: ComposedComponent.displayName,
-    propTypes: {
-      animationDuration: PropTypes.number.isRequired,
-      speed: PropTypes.number
-    },
-    getDefaultProps() {
-      return {
-        animationDuration: defaultAnimationDuration,
-        speed: 1
-      };
-    },
+  class ActivityIndicator extends React.Component {
+    constructor(props) {
+      super(props);
+      this.getDelayStyle = this.getDelayStyle.bind(this);
+      this.getFillStyle = this.getFillStyle.bind(this);
+      this.getBorderStyle = this.getBorderStyle.bind(this);
+    }
     getDelayStyle(delay) {
       const style = {}
       if (delay) {
         style.animationDelay = '-' + (delay * (1 / this.props.speed)) + 's';
       }
       return style;
-    },
+    }
     getFillStyle(delay) {
       const style = this.getDelayStyle(delay);
       if (this.props.color) {
         style.backgroundColor = this.props.color;
       }
       return style;
-    },
+    }
     getBorderStyle(delay) {
       const style = this.getDelayStyle(delay);
       if (this.props.color) {
         style.borderColor = this.props.color;
       }
       return style;
-    },
+    }
     render() {
       const containerStyle = {
         display: 'inline-block',
@@ -50,7 +46,7 @@ export default ((ComposedComponent, defaultAnimationDuration) => {
       let className = 'rai-activity-indicator';
       className += this.props.className ? ` ${this.props.className}` : '';
       return (
-          <div style={containerStyle} className={this.props.className}>
+          <div style={containerStyle} className={className}>
             <ComposedComponent {...this.props}
               getFillStyle={this.getFillStyle}
               getBorderStyle={this.getBorderStyle}
@@ -59,5 +55,19 @@ export default ((ComposedComponent, defaultAnimationDuration) => {
           </div>
       );
     }
-  });
+  }
+
+  ActivityIndicator.propTypes = {
+    animationDuration: PropTypes.number.isRequired,
+    speed: PropTypes.number,
+  };
+
+  ActivityIndicator.defaultProps = {
+    animationDuration: defaultAnimationDuration,
+    speed: 1,
+  };
+
+  ActivityIndicator.displayName = ComposedComponent.name;
+
+  return ActivityIndicator;
 });
