@@ -14,26 +14,23 @@ var _activityIndicator2 = _interopRequireDefault(_activityIndicator);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var Dots = _react2.default.createClass({
-  displayName: 'Dots',
-  render: function render() {
-    return _react2.default.createElement(
-      'div',
-      { style: this.props.style, className: 'rai-dots' },
-      _react2.default.createElement('div', {
-        className: 'rai-circle',
-        style: this.props.getFillStyle(0.3)
-      }),
-      _react2.default.createElement('div', {
-        className: 'rai-circle',
-        style: this.props.getFillStyle(0.2)
-      }),
-      _react2.default.createElement('div', {
-        className: 'rai-circle',
-        style: this.props.getFillStyle(0.1)
-      })
-    );
-  }
-});
+var Dots = function Dots(props) {
+  return _react2.default.createElement(
+    'div',
+    { style: props.style, className: 'rai-dots' },
+    _react2.default.createElement('div', {
+      className: 'rai-circle',
+      style: props.getFillStyle(0.3)
+    }),
+    _react2.default.createElement('div', {
+      className: 'rai-circle',
+      style: props.getFillStyle(0.2)
+    }),
+    _react2.default.createElement('div', {
+      className: 'rai-circle',
+      style: props.getFillStyle(0.1)
+    })
+  );
+};
 
 exports.default = (0, _activityIndicator2.default)(Dots, 0.8);
