@@ -1,15 +1,11 @@
 import React from 'react';
 import activityIndicator from '../activityIndicator';
 
-const Spinner = React.createClass({
-  render() {
-    return (
-      <div style={this.props.style} className="rai-spinner" >
-        <div className="rai-spinner-outer" style={this.props.getBorderStyle()} />
-        <div className="rai-spinner-inner" style={this.props.getBorderStyle()} />
-      </div>
-    );
-  }
-});
+const Spinner = (props) => (
+  <div style={props.style} className="rai-spinner" >
+    <div className="rai-spinner-outer" style={props.getBorderStyle()} />
+    <div className="rai-spinner-inner" style={props.getBorderStyle()} />
+  </div>
+);
 
 export default activityIndicator(Spinner, 0.6);
